# src/config/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
			"URL": {
				"type": "string",
				"example": "https://example.com/example",
				"format": "uri",
				"description": "URL format.",
				"pattern": "^(https?:\\/\\/)?(www\\.)?[-a-zA-Z0-9@:%._\\+~#=]{1,256}\\.[a-zA-Z0-9()]{1,6}\\b([-a-zA-Z0-9()@:%_\\+.~#?&//=]*)$"
			},
...
			"ErrorDetail": {
				"properties": {
					"field": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"value": {
						"type": "string"
					}
				},
				"required": [
					"message"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ForbiddenError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UnauthorizedError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"UUID": {
				"type": "string",
				"example": "52907745-7672-470e-a803-a2f8feb52944",
				"format": "uuid",
				"description": "Stringified UUIDv4.\r\nSee [RFC 4112](https://tools.ietf.org/html/rfc4122)",
				"pattern": "[0-9A-Fa-f]{8}-[0-9A-Fa-f]{4}-4[0-9A-Fa-f]{3}-[89ABab][0-9A-Fa-f]{3}-[0-9A-Fa-f]{12}"
			},
...
			"DefaultString": {
				"type": "string",
				"example": "string",
				"minLength": 2,
				"maxLength": 50
			},
...
			"ConstraintsDatabaseError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Email": {
				"type": "string",
				"example": "example@example.com",
				"format": "email",
				"description": "Email address format.",
				"pattern": "^[a-zA-Z0-9._%+-]+@[a-zA-Z0-9.-]+\\.[a-zA-Z]{2,}$"
			},
			"OrderStatus": {
				"enum": [
					"PENDING",
					"CANCELED",
					"PREPARING",
					"READY",
					"DONE"
				],
				"type": "string"
			},
			"Int": {
				"type": "integer",
				"format": "int32"
			},
			"OrderItemCompleteOut": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/UUID"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"name": {
						"allOf": [
							{
								"$ref": "#/components/schemas/DefaultString"
							}
						],
						"nullable": true
					},
					"pikUrl": {
						"allOf": [
							{
								"$ref": "#/components/schemas/URL"
							}
						],
						"nullable": true
					},
					"price": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"amount": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Int"
							}
						],
						"nullable": true
					},
					"ingredients": {
						"type": "string",
						"nullable": true
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OrderCompleteOut": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/UUID"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"menuId": {
						"allOf": [
							{
								"$ref": "#/components/schemas/UUID"
							}
						],
						"nullable": true
					},
					"customerEmail": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Email"
							}
						],
						"nullable": true
					},
					"totalPrice": {
						"type": "number",
						"format": "double",
						"nullable": true
					},
					"status": {
						"allOf": [
							{
								"$ref": "#/components/schemas/OrderStatus"
							}
						],
						"nullable": true
					},
					"orderItems": {
						"items": {
							"$ref": "#/components/schemas/OrderItemCompleteOut"
						},
						"type": "array",
						"nullable": true
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt",
					"menuId",
					"customerEmail",
					"totalPrice",
					"status",
					"orderItems"
				],
				"type": "object",
				"additionalProperties": false
			},
			"ItemNotFound": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OrderItemCompactIn": {
				"properties": {
					"itemId": {
						"$ref": "#/components/schemas/UUID"
					},
					"amount": {
						"$ref": "#/components/schemas/Int"
					}
				},
				"required": [
					"itemId",
					"amount"
				],
				"type": "object",
				"additionalProperties": false
			},
			"CreateOrderCompactIn": {
				"properties": {
					"items": {
						"items": {
							"$ref": "#/components/schemas/OrderItemCompactIn"
						},
						"type": "array"
					}
				},
				"required": [
					"items"
				],
				"type": "object",
				"additionalProperties": false
			},
...
			"Username": {
				"type": "string",
				"description": "Username format.\r\nMust be 3-30 characters long, containing only letters, numbers, underscores, and hyphens.",
				"pattern": "^[a-zA-Z0-9_-]{3,30}$"
			},
			"IranPhoneNumber": {
				"type": "string",
				"description": "Iran phone number format.\r\nSupported format: 09123456789",
				"pattern": "^0?9\\d{9}$"
			},
...
			"UserCompleteOut": {
				"properties": {
					"id": {
						"$ref": "#/components/schemas/UUID"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"deletedAt": {
						"type": "string",
						"format": "date-time",
						"nullable": true
					},
					"phoneNumber": {
						"allOf": [
							{
								"$ref": "#/components/schemas/IranPhoneNumber"
							}
						],
						"nullable": true
					},
					"username": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Username"
							}
						],
						"nullable": true
					},
					"email": {
						"allOf": [
							{
								"$ref": "#/components/schemas/Email"
							}
						],
						"nullable": true
					}
				},
				"required": [
					"id",
					"createdAt",
					"updatedAt"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserValidationError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"StrongPassword": {
				"type": "string",
				"description": "Strong password format.\r\nMust be 8-64 characters long, containing at least one uppercase letter,\r\none lowercase letter, one number, and one special character (@, $, !, %, *, ?, &).",
				"pattern": "^(?=.*[a-z])(?=.*[A-Z])(?=.*\\d)(?=.*[@$!%*?&])[A-Za-z\\d@$!%*?&]{8,64}$"
			},
			"UserCompactIn": {
				"properties": {
					"phoneNumber": {
						"$ref": "#/components/schemas/IranPhoneNumber"
					},
					"password": {
						"$ref": "#/components/schemas/StrongPassword"
					},
					"username": {
						"$ref": "#/components/schemas/Username"
					},
					"email": {
						"$ref": "#/components/schemas/Email"
					}
				},
				"required": [
					"phoneNumber",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"InvalidCredentialsError": {
				"properties": {
					"name": {
						"type": "string"
					},
					"message": {
						"type": "string"
					},
					"stack": {
						"type": "string"
					},
					"status": {
						"type": "number",
						"format": "double"
					},
					"code": {
						"type": "number",
						"format": "double"
					},
					"details": {
						"items": {
							"$ref": "#/components/schemas/ErrorDetail"
						},
						"type": "array"
					}
				},
				"required": [
					"name",
					"message",
					"status"
				],
				"type": "object",
				"additionalProperties": false
			},
			"UserLogin": {
				"properties": {
					"phoneNumber": {
						"$ref": "#/components/schemas/IranPhoneNumber"
					},
					"password": {
						"type": "string"
					}
				},
				"required": [
					"phoneNumber",
					"password"
				],
				"type": "object",
				"additionalProperties": false
			},
			"SendOtpIn": {
				"properties": {
					"email": {
						"$ref": "#/components/schemas/Email"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": false
			},
			"OTP": {
				"type": "string",
				"example": "12345",
				"format": "otp-5-digit",
				"description": "A 5-digit One-Time Password (OTP).\r\nContains exactly 5 numeric digits (0-9).\r\nMust be a string of length 5.",
				"pattern": "^\\d{5}$",
				"minLength": 5,
				"maxLength": 5
			},
			"CheckOtpIn": {
				"properties": {
					"email": {
						"$ref": "#/components/schemas/Email"
					},
					"code": {
						"$ref": "#/components/schemas/OTP"
					}
				},
				"required": [
					"email",
					"code"
				],
				"type": "object",
				"additionalProperties": false
			}
		},
		"securitySchemes": {
			"SessionAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "session-id"
			},
			"JWTAuth": {
				"type": "apiKey",
				"in": "cookie",
				"name": "access-token"
			}
		}
	},
	"info": {
		"title": "Menuchi-API",
		"version": "1.0.0",
		"description": "Menuchi API Documentation",
		"license": {
			"name": "ISC"
		},
		"contact": {
			"name": "Mohammad Hosseini"
		}
	},
	"paths": {
...
		"/menus/{menuId}/orders": {
			"post": {
				"operationId": "CreateOrder",
				"responses": {
					"201": {
						"description": "Order created successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/OrderCompleteOut"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					},
					"404": {
						"description": "4043 ItemNotFound",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ItemNotFound"
								}
							}
						}
					}
				},
				"description": "Creates a new order for a specific menu.\r\n\r\nAccessible to restaurant customers.",
				"tags": [
					"Order"
				],
				"security": [
					{
						"": [
							"RESTAURANT_CUSTOMER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CreateOrderCompactIn"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetOrders",
				"responses": {
					"200": {
						"description": "All the menu orders retrieved successfully.",
						"content": {
							"application/json": {
								"schema": {
									"items": {
										"$ref": "#/components/schemas/OrderCompleteOut"
									},
									"type": "array"
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Retrieves all orders for a specific menu.",
				"tags": [
					"Order"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Int"
						}
					},
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"$ref": "#/components/schemas/Int"
						}
					},
					{
						"description": "Filters orders based on completion status.\r\n\r\nSet to `false` to retrieve only ongoing (not completed) orders.\r\n\r\nDefaults to `true`, which returns all orders.",
						"in": "query",
						"name": "isCompleted",
						"required": false,
						"schema": {
							"type": "boolean"
						}
					}
				]
			},
			"delete": {
				"operationId": "DeleteOrders",
				"responses": {
					"204": {
						"description": "Orders deleted successfully. It doesn't retrieve anything.",
						"content": {
							"application/json": {
								"schema": {
									"type": "number",
									"enum": [
										null
									],
									"nullable": true
								}
							}
						}
					},
					"401": {
						"description": "Unauthorized user.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UnauthorizedError"
								}
							}
						}
					},
					"403": {
						"description": "Access Denied. You are not authorized to perform this action.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ForbiddenError"
								}
							}
						}
					}
				},
				"description": "Deletes multiple orders from a menu.",
				"tags": [
					"Order"
				],
				"security": [
					{
						"": [
							"RESTAURANT_OWNER"
						]
					}
				],
				"parameters": [
					{
						"in": "path",
						"name": "menuId",
						"required": true,
						"schema": {
							"$ref": "#/components/schemas/UUID"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"items": {
									"$ref": "#/components/schemas/UUID"
								},
								"type": "array"
							}
						}
					}
				}
			}
		},
...
		"/auth/res-signup": {
			"post": {
				"operationId": "RestaurantOwnerSignup",
				"responses": {
					"201": {
						"description": "User signed up successfully.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserCompleteOut"
								}
							}
						}
					},
					"409": {
						"description": "ConstraintsDatabaseError -> A user with the provided credentials already exists.",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ConstraintsDatabaseError"
								}
							}
						}
					},
					"422": {
						"description": "4225 UserValidationError",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/UserValidationError"
								}
							}
						}
					}
				},
				"description": "Registers a new restaurant owner.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserCompactIn"
							}
						}
					}
				}
			}
		},
		"/auth/res-signin": {
			"post": {
				"operationId": "RestaurantOwnerSignin",
				"responses": {
					"200": {
						"description": "User signed in successfully.",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					},
					"401": {
						"description": "InvalidCredentialsError",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/InvalidCredentialsError"
								}
							}
						}
					}
				},
				"description": "Authenticates a restaurant owner.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/UserLogin"
							}
						}
					}
				}
			}
		},
		"/auth/send-otp": {
			"post": {
				"operationId": "SendOtp",
				"responses": {
					"200": {
						"description": "Otp code sent successfully.",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"description": "Sends an OTP to the email.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendOtpIn"
							}
						}
					}
				}
			}
		},
		"/auth/check-otp": {
			"post": {
				"operationId": "CheckOtp",
				"responses": {
					"200": {
						"description": "User authenticated successfully.",
						"content": {
							"application/json": {
								"schema": {
									"type": "boolean"
								}
							}
						}
					}
				},
				"description": "Sends email and OTP for auth.",
				"tags": [
					"Auth"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/CheckOtpIn"
							}
						}
					}
				}
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	],
...
}